python_non_api/pptx_generator/requirements.txt
python-pptx==0.6.23
pyyaml